Bug fix #169: Display user log date in numeric format
Also rename Timezone to Zone to avoid waste of space in page.

+++ ui/js/src/wok.utils.js
@@ -244,7 +244,7 @@ wok.datetimeLocaleConverter = function datetimeLocaleConverter(datetime_string, 
 
 wok.dateLocaleConverter = function dateLocaleConverter(date_string, locale){
      var dte = new moment(date_string);
-     var options = { year: 'numeric', month: 'long', day: 'numeric' };
+     var options = { year: 'numeric', month: 'numeric', day: 'numeric' };
      return dte._d.toLocaleDateString(locale, options);
 }
